==> 04_exploratory_analysis.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r message=FALSE}
library(tidyverse)
library(timetk)

appointments_tbl <- read_rds("00_data/processed/wakefield_weekdays_full.rds")
```

# 1: Summary Statistics

```{r}
skimr::skim(appointments_tbl)
```

-   1391 days between 2019-03-01 and 2024-06-28

-   numeric response variable with no missing values

-   4 categorical variables (nominal)

# 2: Exploratory Visualisation

## 2a: Distribution of appointments

```{r}
# overall
total_daily_appointments <- 
    appointments_tbl %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        total_appointments = sum(count_of_appointments)
    )

total_daily_appointments %>% 
    ggplot(aes(total_appointments)) +
    geom_histogram(bins = 20, fill = "skyblue", color = "grey30") +
    theme_bw() +
    labs(title = "Distribution of Total Daily Appointments")

total_daily_appointments %>% 
    filter(total_appointments > 0) %>%  
    select(total_appointments) %>% 
    summary()
```

```{r}
gp_daily_appointments <- 
    appointments_tbl %>%
    filter(hcp_type == "GP") %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        total_appointments = sum(count_of_appointments)
    )

gp_daily_appointments %>% 
    ggplot(aes(total_appointments)) +
    geom_histogram(bins = 20, fill = "skyblue", color = "grey30") +
    theme_bw() +
    labs(title = "Distribution of GP Daily Appointments")

gp_daily_appointments %>% 
    filter(total_appointments > 0) %>%  
    select(total_appointments) %>% 
    summary()
```

```{r}
other_staff_daily_appointments <- 
    appointments_tbl %>%
    filter(hcp_type == "Other Practice staff") %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        total_appointments = sum(count_of_appointments)
    )

other_staff_daily_appointments %>% 
    ggplot(aes(total_appointments)) +
    geom_histogram(bins = 20, fill = "skyblue", color = "grey30") +
    theme_bw() +
    labs(title = "Distribution of Other Practice staff Daily Appointments")

other_staff_daily_appointments %>% 
    filter(total_appointments > 0) %>%  
    select(total_appointments) %>% 
    summary()
```

## \*\* FINDINGS

-   approximately normal distributions when values for Bank Holidays omitted

## 2b: Total Appointments by Feature

```{r}
appointments_tbl %>%
    group_by(hcp_type) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    mutate(hcp_type = fct_reorder(hcp_type, -total_appointments)) %>%
    ggplot(aes(hcp_type, total_appointments)) +
    geom_col(fill = "skyblue", color = "grey30") +
    theme_bw() +
    labs(title = "Total Appointments by HCP Type")
```

```{r}
appointments_tbl %>%
    group_by(appt_mode) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    mutate(appt_mode = fct_reorder(appt_mode, -total_appointments)) %>%
    ggplot(aes(appt_mode, total_appointments)) +
    geom_col(fill = "skyblue", color = "grey30") +
    theme_bw() +
    labs(title = "Total Appointments by Appointment Mode")
```

```{r}
appointments_tbl %>%
    group_by(appt_status) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    mutate(appt_status = fct_reorder(appt_status, -total_appointments)) %>%
    ggplot(aes(appt_status, total_appointments)) +
    geom_col(fill = "skyblue", color = "grey30") +
    theme_bw() +
    labs(title = "Total Appointments by Appointment Status")
```

```{r}
appointments_tbl %>%
    group_by(time_between_book_and_appt) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    ggplot(aes(time_between_book_and_appt, total_appointments)) +
    geom_col(fill = "skyblue", color = "grey30") +
    theme_bw() +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5)) +
    labs(title = "Total Appointments by Time Between Booking and Appointment Date",
         x = "Time between booking and appointment date")

appointments_tbl %>%
    mutate(
        booking_type = if_else(
            time_between_book_and_appt == "Same Day",
            "Same Day",
            "Advance"
        )
    ) %>%
    group_by(booking_type) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    ggplot(aes(booking_type, total_appointments)) +
    geom_col(fill = "skyblue", color = "grey30") +
    theme_bw() +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5)) +
    labs(title = "Total Appointments by Booking Type",
         x = "Booking Type")
```

## \*\* FINDINGS

-   More appointments with Other Practice Staff than with GPs

-   More appointments Face-to-Face than other modes

-   Only a small percentage DNAs (but not sure what 'unknown' means)

-   Same day most frequent category of time between booking and appointment date

    -   Overall more appointments advance booked where all non same day appointments aggregated

## 2c: Appointments vs. HCP Type by Feature

```{r}
appointments_tbl %>%
    filter(hcp_type == "GP" | hcp_type == "Other Practice staff",
           appt_mode %in% c("Face-to-Face", "Telephone")) %>% 
    group_by(hcp_type, appt_mode) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    mutate(appt_mode = fct_reorder(appt_mode, total_appointments)) %>%
    ggplot(aes(hcp_type, total_appointments, fill = appt_mode)) +
    geom_col(color = "grey30", position = "fill") +
    theme_bw() +
    tidyquant::scale_fill_tq()
```

```{r}
appointments_tbl %>%
    filter(hcp_type == "GP" | hcp_type == "Other Practice staff",
           appt_mode %in% c("Face-to-Face", "Telephone")) %>% 
    group_by(hcp_type, appt_status) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    mutate(appt_status = fct_reorder(appt_status, total_appointments)) %>%
    ggplot(aes(hcp_type, total_appointments, fill = appt_status)) +
    geom_col(color = "grey30", position = "fill") +
    theme_bw() +
    tidyquant::scale_fill_tq()
```

```{r}
appointments_tbl %>%
    filter(hcp_type == "GP" | hcp_type == "Other Practice staff") %>% 
    group_by(hcp_type, time_between_book_and_appt) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    mutate(prop_appointment = proportions(total_appointments)) %>% 
    ggplot(aes(hcp_type, prop_appointment, fill = time_between_book_and_appt)) +
    geom_col(color = "grey30", position = "dodge") +
    theme_bw() +
    tidyquant::scale_fill_tq()

appointments_tbl %>%
    filter(hcp_type == "GP" | hcp_type == "Other Practice staff") %>% 
    group_by(hcp_type, time_between_book_and_appt) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    mutate(hcp_type = fct_rev(hcp_type)) %>% 
    ggplot(aes(time_between_book_and_appt, total_appointments, fill = hcp_type)) +
    geom_col(position = "fill") +
    theme_bw() +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5)) +
    xlab("Time between booking and appointment date") +
    tidyquant::scale_fill_tq()

appointments_tbl %>%
    filter(hcp_type == "GP" |
               hcp_type == "Other Practice staff") %>%
    mutate(booking_type = if_else(time_between_book_and_appt == "Same Day",
                                  "Same Day",
                                  "Advance")) %>%
    group_by(hcp_type, booking_type) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    mutate(hcp_type = hcp_type) %>%
    ggplot(aes(hcp_type, total_appointments, fill = booking_type)) +
    geom_col(position = "fill") +
    theme_bw() +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5)) +
    xlab("Time between booking and appointment date") +
    tidyquant::scale_fill_tq()
```

```{r}
appointments_tbl %>%
    filter(hcp_type == "GP" | hcp_type == "Other Practice staff") %>%   
    mutate(appt_status = fct_relevel(appt_status, c("Attended", "Unknown", "DNA"))) %>%
    group_by(hcp_type, appt_status, time_between_book_and_appt) %>%
    summarise(total_appointments = sum(count_of_appointments)) %>%
    ggplot(aes(time_between_book_and_appt, total_appointments, fill = appt_status)) +
    geom_col(position = "fill") +
    theme_bw() +
    tidyquant::scale_fill_tq() +
    labs(title = "Very few DNA or Unknown status for same day appointments",
         subtitle = "DNA and Unknown proportion increases as delay increases") +
    facet_wrap(~ hcp_type) +
    theme(axis.text.x = element_text(angle = 90, hjust = 1))
    
```

## \*\* FINDINGS

-   Proportion of face to face appointments higher for other staff compared to GPs

-   Proportion of DNA/Unknown appointments higher for other staff compared to GPs

-   Proportion of same day appointments higher for GPs compared to other staff

-   Lower DNA rate for same day appointments - rate increases with time between booking and appointment

# 3: Time Series Analysis

Focus on appointments recorded as attended. This provides an overview of the workload.

Look at total appointments plus subgroups - GP vs. Other practice staff, telephone vs. face to face, same day vs advanced booking.Finally look at all combinations.

## 3a: Total Appointments

```{r}
appointments_filtered_tbl <-
    appointments_tbl %>%
    filter(
        hcp_type == "GP" | hcp_type == "Other Practice staff",
        appt_status == "Attended",
        appt_mode %in% c("Face-to-Face", "Telephone"),
        time_between_book_and_appt != "Unknown"
    ) %>%
    mutate(booking_type = if_else(time_between_book_and_appt == "Same Day",
                                  "Same Day",
                                  "Advance"))
```

```{r}
appointments_filtered_tbl %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        appointments = sum(count_of_appointments)
    ) %>%
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Total Daily Appointments")
```

```{r}
appointments_filtered_tbl %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "week",
        appointments = sum(count_of_appointments)
    ) %>%
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Total Weekly Appointments")
```

```{r}
appointments_filtered_tbl %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "month",
        appointments = sum(count_of_appointments)
    ) %>%
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Total Monthly Appointments")
```

## \*\* FINDINGS

-   significant fall in appointments in early 2020 due to pandemic

-   from mid 2020 to mid 2023 appointments increased

-   since late 2023 total appointments plateaued.

## 3b: Total Appointments by HCP Type

```{r}
total_appts_by_hcp <-
    appointments_filtered_tbl %>% 
    group_by(hcp_type) %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        appointments = sum(count_of_appointments))
```

```{r}
# by day
total_appts_by_hcp %>% 
    group_by(hcp_type) %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        appointments = sum(appointments)) %>% 
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Daily Appointments by HCP Type")
```

```{r}
# by week
total_appts_by_hcp %>% 
    group_by(hcp_type) %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "week",
        appointments = sum(appointments)) %>% 
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Weekly Appointments by HCP Type")
```

```{r}
# by month
total_appts_by_hcp %>% 
    group_by(hcp_type) %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "month",
        appointments = sum(appointments)) %>% 
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Monthly Appointments by HCP Type")
```

## \*\* FINDINGS

-   GP and Other Practice staff appointments both fell due to pandemic

-   From mid 2020 GP appointments increased up to peak in early 2022 but since then rate falling

-   From early 2021 Other practice staff appointments have been increasing

-   It appears that plateau in total appointments consists of a fall in GP appointments combined with an increase in appointments with other practice staff

## 3c: Total Appointments by Booking Type

```{r}
total_appts_by_booking_type <-
    appointments_filtered_tbl %>% 
    group_by(booking_type) %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        appointments = sum(count_of_appointments))
```

```{r}
# by day
total_appts_by_booking_type %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        appointments = sum(appointments)) %>% 
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Daily Appointments by Booking Type")
```

```{r}
# by week
total_appts_by_booking_type %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "week",
        appointments = sum(appointments)) %>% 
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Weekly Appointments by Booking Type")
```

```{r}
# by month
total_appts_by_booking_type %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "month",
        appointments = sum(appointments)) %>% 
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Monthly Appointments by Booking Type")
```

## \*\* FINDINGS

-   advance bookings fell significantly in early 2020 due to the pandemic but same day bookings did not seem to be impacted

-   advance and same day appointments bookings have been increasing since but since the peak in late 2022/early 2023 same day appointments appear to be decreasing

## 3d: GP Appointments by Mode

```{r}
total_appts_by_appt_mode <-
    appointments_filtered_tbl %>% 
    group_by(appt_mode) %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        appointments = sum(count_of_appointments))
```

```{r}
# by day
total_appts_by_appt_mode %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        appointments = sum(appointments)) %>% 
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Daily Appointments by Mode")
```

```{r}
# by week
total_appts_by_appt_mode %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "week",
        appointments = sum(appointments)) %>% 
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Weekly Appointments by Mode")

```

```{r}
# by month
total_appts_by_appt_mode %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "month",
        appointments = sum(appointments)) %>% 
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .title = "Time Series Plot: Monthly Appointments by Mode")
```

## \*\* FINDINGS

-   face to face and telephone appointments appear to be inversely related

-   in 2020, due to the pandemic, there appeared to be a shift from face to face to telephone appointments

-   from early 2021 telephone appointments have been falling whilst face to face have increased

-   both appear to be plateauing since start of 2024

# 4: Time Series Analysis - Sub Groups

## 4a: Compare all subgroups

```{r}
total_appts_by_subgroup <-
    appointments_filtered_tbl %>% 
    group_by(appt_mode, hcp_type, booking_type) %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        appointments = sum(count_of_appointments))
```

```{r}
# by month
total_appts_by_subgroup %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "week",
        appointments = sum(appointments)) %>% 
    plot_time_series(.date_var = appointment_date,
                     .value = appointments,
                     .trelliscope = TRUE,
                     .facet_strip_remove = TRUE)
```

## \*\* FINDINGS

-   face to face appointments for all subgroups of hcp and booking type demonstrate similar pattern

    -   fell during early pandemic then increased until early 2024 then plateaued

-   telephone appointments for GPs increase during pandemic for both advance booked and same day appointments.

    -   fallen since mid 2021 - decrease continuing although potential plateauing

-   telephone appointments for other practice staff shows different pattern

    -   same day appointments showed increase in pandemic but now back to prepandemic levels and stable

    -   advance booked appointments appear to increasingly be delivered by telephone

        -   is this non urgent / planned care or triage??

            -   would need to analyse data with original booking type levels to explore this further

-   EACH GROUP APPEARS TO SHOW SIGNIFICANT IMPACT OF PANDEMIC

    -   RE-VISUALISE POST PANDEMIC (2021-08-01 +)

-   EACH GROUP SMOOTHED CURVE APPEARS TO HAVE BEEN INFLUENCED BY POTENTIAL ANOMALOUS VALUES LATE 2022 EARLY 2023

## 4b: GP Face-to-Face

```{r}
gp_f2f_weekly <- 
    total_appts_by_subgroup %>% 
    filter(appt_mode == "Face-to-Face", hcp_type == "GP") %>% 
    group_by(booking_type) %>% 
    summarise_by_time(.date_var = appointment_date, .by = "week", appointments = sum(appointments))
```

```{r}
gp_f2f_weekly %>%
    plot_time_series(
        .date_var = appointment_date,
        .value = appointments,
        .smooth_period = 65,
        .smooth_degree = 2,
        .title = "GP Face-to-Face: Weekly Time Series"
    )
```

```{r}
gp_f2f_weekly %>%
    filter_by_time(.date_var = appointment_date, .start_date = "2021-08-01") %>%
    plot_time_series(
        .date_var = appointment_date,
        .value = appointments,
        .smooth_period = 260,
        .smooth_degree = 2, 
        .title = "GP Face-to-Face Post Pandemic: Weekly Time Series"
    )
```

```{r}
gp_f2f_weekly %>%
    filter_by_time(.date_var = appointment_date, .start_date = "2021-08-01") %>%
    plot_anomaly_diagnostics(
        .date_var = appointment_date,
        .value = appointments,
        .alpha = 0.05,
        .interactive = TRUE,
        .frequency = 52,
        .title = "GP Face-to-Face Post Pandemic: Weekly Anomalies"
    )
```

## \*\* FINDINGS

## 4c: GP Telephone

```{r}
gp_telephone_weekly <- 
    total_appts_by_subgroup %>% 
    filter(appt_mode == "Telephone", hcp_type == "GP") %>% 
    group_by(booking_type) %>% 
    summarise_by_time(.date_var = appointment_date, .by = "week", appointments = sum(appointments))
```

```{r}
gp_telephone_weekly %>%
    plot_time_series(
        .date_var = appointment_date,
        .value = appointments,
        .smooth_period = 65,
        .smooth_degree = 2,
        .title = "GP Telephone: Weekly Time Series"
    )
```

```{r}
gp_telephone_weekly %>%
    filter_by_time(.date_var = appointment_date, .start_date = "2021-08-01") %>%
    plot_time_series(
        .date_var = appointment_date,
        .value = appointments,
        .smooth_period = 260,
        .smooth_degree = 2, 
        .title = "GP Telephone Post Pandemic: Weekly Time Series"
    )
```

```{r}
gp_telephone_weekly %>%
    filter_by_time(.date_var = appointment_date, .start_date = "2021-08-01") %>%
    plot_anomaly_diagnostics(
        .date_var = appointment_date,
        .value = appointments,
        .alpha = 0.05,
        .interactive = TRUE,
        .frequency = 52,
        .title = "GP Telephone Post Pandemic: Weekly Anomalies"
    )
```

## \*\* FINDINGS

## 4d: Proportion Same Day vs Advance Booked over Time

```{r}
gp_appts_percent_sameday_weekly <- 
    total_appts_by_subgroup %>%
    filter(hcp_type == "GP") %>%
    group_by(booking_type) %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "week",
        appointments = sum(appointments)
    ) %>% 
    pivot_wider(names_from = booking_type, values_from = appointments) %>% 
    mutate(same_day_proportion = `Same Day` / (`Same Day` + Advance),
           same_day_percent = round(same_day_proportion * 100, 2)) %>% 
    select(appointment_date, same_day_percent)
```

```{r}
gp_appts_percent_sameday_weekly %>% 
    plot_time_series(.date_var = appointment_date, .value = same_day_percent)
```

```{r}
gp_appts_percent_sameday_post_pandemic_weekly <- 
    gp_appts_percent_sameday_weekly %>% 
    filter_by_time(.date_var = appointment_date, .start_date = "2021-08-01")
    
gp_appts_percent_sameday_post_pandemic_weekly %>% 
    plot_time_series(.date_var = appointment_date, .value = same_day_percent, 
                     .y_lab = "Percent Same Day",
                     .title = "Percent GP Appts booked on Same Day Post-Pandemic: Weekly")
```

```{r}
gp_appts_percent_sameday_post_pandemic_weekly %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "month",
        same_day_percent = mean(same_day_percent)
    ) %>%
    plot_time_series(
        .date_var = appointment_date,
        .value = same_day_percent,
        .y_lab = "Percent Same Day",
        .title = "Percent GP Appts booked on Same Day Post-Pandemic: Monthly"
    )
```

## \*\* FINDINGS

NOT FINISHED SECTIONS BELOW HERE

# 5: Autocorrelation analysis

I am most interested in Same Day GP appointments broken down into total, face-to-face. and telephone.

## Total Same Day GP Appointments

```{r}
gp_appt_mode_vs_booking_tbl %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        total_appointments = sum(appointments)
    ) %>% 
    plot_time_series(.date_var = appointment_date, .value = total_appointments)

gp_appt_mode_vs_booking_tbl %>%
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        total_appointments = sum(appointments)
    ) %>%
    plot_acf_diagnostics(.date_var = appointment_date, .value = total_appointments, .lags = 365)
    
```

## Same Day GP Appointments by Appointment Mode

```{r}
gp_appt_mode_vs_booking_tbl %>%
    group_by(appt_mode) %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        total_appointments = sum(appointments)
    ) %>% 
    plot_time_series(.date_var = appointment_date, .value = total_appointments)
```

```{r}
gp_appt_mode_vs_booking_tbl %>%
    group_by(appt_mode) %>% 
    summarise_by_time(
        .date_var = appointment_date,
        .by = "day",
        total_appointments = sum(appointments)
    ) %>%
    plot_acf_diagnostics(.date_var = appointment_date, .value = total_appointments, .lags = 365)
    
```

### Findings:

# Exploring Seasonality

Focus on GP same day data - Totals vs Phone vs Face to face

```{r}

gp_appt_mode_vs_booking_tbl %>% 
    pivot_wider(names_from = appt_mode, values_from = appointments) %>% 
    mutate(Total = `Face-to-Face` + Telephone) %>% 
    pivot_longer(cols = 3:5, names_to = "appt_mode", values_to = "appointments")
```

```{r}
gp_appt_mode_vs_booking_tbl %>% 
    group_by(appt_mode) %>% 
    plot_seasonal_diagnostics(
        .date_var = appointment_date,
        .value = appointments,
        .feature_set = c("wday.lbl", "month.lbl", "year"),
        .title = ""
    )
```

```{r}
same_day_vs_booked_daily_tbl %>%
    filter(!appt_mode %in% c("Unknown", "Video Conference/Online")) %>%
    plot_seasonal_diagnostics(
        .date_var = appointment_date,
        .value = appointments,
        .feature_set = c("wday.lbl", "month.lbl", "year"),
        .title = "",
        .facet_vars = appt_mode,
        .interactive = FALSE
    ) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5))

wakefield_working_week_daily %>%
    filter(hcp_type == "GP" | hcp_type == "Other Practice staff",
           appt_mode %in% c("Face-to-Face", "Telephone")) %>% 
    mutate(day = wday(appointment_date, label = TRUE),
           appt_mode = fct_rev(appt_mode)) %>% 
    group_by(day, hcp_type, appt_mode) %>% 
    summarise(total_appointments = sum(count_of_appointments)) %>%
    ungroup() %>% 
    ggplot(aes(day, total_appointments, fill = appt_mode)) +
    geom_col(position = "fill") +
    theme_bw() +
    facet_wrap(~ hcp_type) +
    scale_fill_brewer(type = "qual", palette = 3)

wakefield_working_week_daily %>%
    filter(hcp_type == "GP" | hcp_type == "Other Practice staff",
           appt_mode %in% c("Face-to-Face", "Telephone")) %>% 
    mutate(month = month(appointment_date, label = TRUE),
           appt_mode = fct_rev(appt_mode)) %>% 
    group_by(month, hcp_type, appt_mode) %>% 
    summarise(total_appointments = sum(count_of_appointments)) %>%
    ungroup() %>% 
    ggplot(aes(month, total_appointments, fill = appt_mode)) +
    geom_col(position = "fill") +
    theme_bw() +
    facet_wrap(~ hcp_type) +
    scale_fill_brewer(type = "qual", palette = 3)

same_day_vs_booked_daily_tbl %>%
    filter(!appt_mode %in% c("Unknown", "Video Conference/Online")) %>%
    plot_seasonal_diagnostics(
        .date_var = appointment_date,
        .value = appointments,
        .feature_set = c("wday.lbl", "month.lbl", "year"),
        .title = "",
        .facet_vars = c(booking, hcp_type),
        .interactive = FALSE,
    ) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5))

same_day_vs_booked_daily_tbl %>% 
    mutate(day = wday(appointment_date, label = TRUE)) %>% 
    group_by(day, hcp_type, booking) %>% 
    summarise(appointments = sum(appointments)) %>%
    ungroup() %>% 
    ggplot(aes(day, appointments, fill = booking)) +
    geom_col(position = "fill") +
    theme_bw() +
    facet_wrap(~ hcp_type) +
    scale_fill_brewer(palette = 3)

same_day_vs_booked_daily_tbl %>% 
    mutate(month = month(appointment_date, label = TRUE),
           appt_mode = fct_rev(appt_mode)) %>% 
    group_by(month, hcp_type, booking) %>%
    summarise(appointments = sum(appointments)) %>%
    ungroup() %>% 
    ggplot(aes(month, appointments, fill = booking)) +
    geom_col(position = "fill") +
    theme_bw() +
    facet_wrap(~ hcp_type) +
    scale_fill_brewer(type = "qual", palette = 3)

```

## Correlation

```{r}
total_appts_by_subgroup %>% 
    pivot_wider(id_cols = appointment_date, names_from = c(hcp_type, appt_mode, booking_type), values_from = appointments) %>%
    select(-appointment_date) %>% 
    cor() %>% 
    corrplot::corrplot()
    
```

```{r}
attended_daily_tbl %>% 
    filter(hcp_type != "Total", appointments > 0) %>%
    pivot_wider(names_from = hcp_type, values_from = appointments) %>% 
    ggplot(aes(GP, `Other Practice staff`)) +
    geom_point(alpha = 0.5) +
    geom_smooth(method = 'lm') +
    theme_bw()

attended_daily_tbl %>% 
    filter(hcp_type != "Total", appointments > 0) %>%
    pivot_wider(names_from = hcp_type, values_from = appointments) %>% 
    select(2:3) %>% 
    cor()
```

Total GP appointments attended per day moderately strongly correlated to total Other Practice staff appointments (when zero values omitted)

GP workload is simpler to analyse because in this dataset "Other Practice Staff" is a homogeneous group that may contain HCA, PA, Nurse, Pharmacists etc. and may be better utilised as a feature rather than a response

GP appointments fell around pandemic then increased to a peak between 2022 and 2023. Values appear to be falling since 2023 . Total appointments increasing so reduced total gp workload occurring as workload for other practice staff increasing.

A further question to consider relates to workforce. What are appointment rates per GP? Can workforce data be used to estimate this? Is there any data that describes daily GP sessions?
